reorganize parser
